chore(testing): add shared base Vitest configuration
- Added config/vitest.base.mjs providing createBaseConfig helper using import.meta.url
- Lifted vitest and @vitest/browser-playwright dependencies to root package.json devDependencies
- Refactored packages/logger/vitest.config.mjs to use createBaseConfig(import.meta.url)
- Centralized Playwright browser initialization, concurrency limits, dynamic setupFiles, and Chai/Sinon pre-bundling
