+++ package.json
@@ -86,6 +86,7 @@
     "@typescript-eslint/eslint-plugin": "7.18.0",
     "@typescript-eslint/eslint-plugin-tslint": "7.0.2",
     "@typescript-eslint/parser": "7.18.0",
+    "@vitest/browser-playwright": "4.1.10",
     "babel-loader": "10.1.1",
     "babel-plugin-istanbul": "8.0.0",
     "chai": "4.5.0",
@@ -140,6 +141,7 @@
     "tslint": "6.1.3",
     "typedoc": "0.16.11",
     "typescript": "5.5.4",
+    "vitest": "4.1.10",
     "webpack": "5.104.1",
     "yargs": "17.7.2"
   }
